restructured constants folder
