+++ packages/web-app/src/utils/constants/api.ts
@@ -1,0 +1,32 @@
+type SubgraphNetworkUrl = {
+  [key: string]: string | undefined;
+};
+
+export const SUBGRAPH_API_URL: SubgraphNetworkUrl = {
+  main: undefined,
+  // rinkeby:
+  //   'https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-rinkeby',
+  rinkeby:
+    'https://api.thegraph.com/subgraphs/name/rekard0/rekard0-zaragoza-rinkeby',
+  polygon: undefined,
+  mumbai:
+    'https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-mumbai',
+  arbitrum: undefined,
+  arbitrumTest:
+    'https://api.thegraph.com/subgraphs/name/aragon/aragon-zaragoza-arbitrum-rinkeby',
+};
+
+export const BASE_URL = 'https://api.coingecko.com/api/v3';
+export const DEFAULT_CURRENCY = 'usd';
+export const INFURA_PROJECT_ID = '7a03fcb37be7479da06f92c5117afd47';
+export const INFURA_PROJECT_ID_ARB = '92aa62d2bb5449cfafe04b83ca8636f1';
+
+// Coingecko Api specific asset platform keys
+export const ASSET_PLATFORMS: {[key: number]: string} = {
+  1: 'ethereum',
+  127: 'polygon-pos',
+  42161: 'arbitrum-one',
+};
+
+// to be removed
+export const TEST_DAO = '0x51c3ddb42529bfc24d4c13192e2e31421de459bc';
