+++ packages/web-app/src/utils/constants/misc.ts
@@ -1,0 +1,42 @@
+import {
+  IconCommunity,
+  IconDashboard,
+  IconFinance,
+  IconGovernance,
+} from '@aragon/ui-components';
+
+import {i18n} from '../../../i18n.config';
+import {Dashboard, Community, Finance, Governance} from '../paths';
+
+/** Time period options for token price change */
+export const enum TimeFilter {
+  day = 'day',
+  week = 'week',
+  month = 'month',
+  year = 'year',
+  // max = 'max',
+}
+
+export const NAV_LINKS = [
+  {
+    label: i18n.t('navLinks.dashboard'),
+    path: Dashboard,
+    icon: IconDashboard,
+  },
+  {
+    label: i18n.t('navLinks.governance'),
+    path: Governance,
+    icon: IconGovernance,
+  },
+  {label: i18n.t('navLinks.finance'), path: Finance, icon: IconFinance},
+  {
+    label: i18n.t('navLinks.community'),
+    path: Community,
+    icon: IconCommunity,
+  },
+];
+
+export const enum TransferTypes {
+  Deposit = 'Deposit',
+  Withdraw = 'Withdraw',
+}
